requirements.txt
pandas
transformers
tensorflow-cpu
gpt_2_simple